delete branch when existing

+++ entrypoint.sh
@@ -51,7 +51,10 @@ cd "$CLONE_DIR"
 if [ ! -z "$INPUT_DESTINATION_BRANCH_CREATE" ]
 then
   echo "Creating new branch: ${INPUT_DESTINATION_BRANCH_CREATE}"
-  git push origin --delete "$INPUT_DESTINATION_BRANCH_CREATE"
+  exists=$(git ls-remote --heads origin ${INPUT_DESTINATION_BRANCH_CREATE})
+  if [[ -z ${exists} ]]; then
+    git push origin --delete "$INPUT_DESTINATION_BRANCH_CREATE"
+  else
   git checkout -b "$INPUT_DESTINATION_BRANCH_CREATE"
   OUTPUT_BRANCH="$INPUT_DESTINATION_BRANCH_CREATE"
 fi
